Create README.md
Base of a framework for testing Windows security for PCI-DSS 3. 
Lots of cheating e.g. Powershell scripts are on the target device, so this is not exactly a remote testing solution as yet, but the first steps, with a default Django templates, sets the way forward for this project.

Note the requirements for your virtualenv in requirements.txt.